Remove check on number of profile data
The check of number of data in NemoFeedbackReduce::reduce_profile_data incorrectly throws an error when a whole profile is not being written to file. This is because data_in does not include the information about those profiles because they are ignored in https://github.com/MetOffice/nemo-feedback/blob/develop/src/nemo-feedback/NemoFeedback.cc#L448.